+++ crates/ryft-xla-sys/src/mlir/dialects.rs
@@ -1,4 +1,5 @@
 pub mod affine;
+pub mod arith;
 pub mod gpu;
 pub mod mosaic;
 pub mod triton;